requirements.txt
Django==2.2
pytz==2020.1
sqlparse==0.3.1
PyJWT==1.7.1
psycopg2-binary==2.8.5
python-memcached==1.59
requests==2.24.0